+++ src/Sentry/Internal/Json.cs
@@ -26,10 +26,20 @@ namespace Sentry.Internal
             new StreamWriter(stream, Encoding, 1024, true)
         );
 
+        private static JsonTextReader CreateReader(Stream stream) => new JsonTextReader(
+            new StreamReader(stream, Encoding, false, 1024, true)
+        );
+
         public static void SerializeToStream(object obj, Stream stream)
         {
             using var writer = CreateWriter(stream);
             Serializer.Serialize(writer, obj);
+        }
+
+        public static T DeserializeFromStream<T>(Stream stream)
+        {
+            using var reader = CreateReader(stream);
+            return Serializer.Deserialize<T>(reader);
         }
 
         public static byte[] SerializeToByteArray(object obj)
@@ -40,9 +50,17 @@ namespace Sentry.Internal
             return buffer.ToArray();
         }
 
-        public static string Serialize(object obj) => Encoding.GetString(
-            SerializeToByteArray(obj)
-        );
+        public static T DeserializeFromByteArray<T>(byte[] data)
+        {
+            using var buffer = new MemoryStream(data);
+            return DeserializeFromStream<T>(buffer);
+        }
+
+        public static string Serialize(object obj) =>
+            Encoding.GetString(SerializeToByteArray(obj));
+
+        public static T Deserialize<T>(string json) =>
+            DeserializeFromByteArray<T>(Encoding.GetBytes(json));
 
         public static async Task SerializeToStreamAsync(
             object obj,