+++ src/Sentry/Protocol/Envelope.cs
@@ -4,13 +4,14 @@ using System.IO;
 using System.Threading;
 using System.Threading.Tasks;
 using Sentry.Internal;
+using Sentry.Internal.Extensions;
 
 namespace Sentry.Protocol
 {
     /// <summary>
     /// Envelope.
     /// </summary>
-    public class Envelope : ISerializable
+    public class Envelope : IDisposable, ISerializable
     {
         private const string EventIdKey = "event_id";
 
@@ -48,15 +49,18 @@ namespace Sentry.Protocol
         {
             // Header
             await Json.SerializeToStreamAsync(Header, stream, cancellationToken).ConfigureAwait(false);
-            stream.WriteByte((byte)'\n');
+            await stream.WriteByteAsync((byte)'\n', cancellationToken).ConfigureAwait(false);
 
             // Items
             foreach (var item in Items)
             {
                 await item.SerializeAsync(stream, cancellationToken).ConfigureAwait(false);
-                stream.WriteByte((byte)'\n');
+                await stream.WriteByteAsync((byte)'\n', cancellationToken).ConfigureAwait(false);
             }
         }
+
+        /// <inheritdoc />
+        public void Dispose() => Items.DisposeAll();
 
         /// <summary>
         /// Creates an envelope that contains a single event.
@@ -90,6 +94,49 @@ namespace Sentry.Protocol
             {
                 EnvelopeItem.FromUserFeedback(sentryUserFeedback)
             };
+
+            return new Envelope(header, items);
+        }
+
+        private static async Task<IReadOnlyDictionary<string, object>> DeserializeHeaderAsync(
+            Stream stream,
+            CancellationToken cancellationToken = default)
+        {
+            var buffer = new List<byte>();
+
+            var prevByte = default(int);
+            await foreach (var curByte in stream.ReadAllBytesAsync(cancellationToken))
+            {
+                // Break if found an unescaped newline
+                if (curByte == '\n' && prevByte != '\\')
+                {
+                    break;
+                }
+
+                buffer.Add(curByte);
+                prevByte = curByte;
+            }
+
+            return
+                Json.DeserializeFromByteArray<Dictionary<string, object>?>(buffer.ToArray())
+                ?? throw new InvalidOperationException("Envelope header is malformed.");
+        }
+
+        /// <summary>
+        /// Deserializes envelope from stream.
+        /// </summary>
+        public static async Task<Envelope> DeserializeAsync(
+            Stream stream,
+            CancellationToken cancellationToken = default)
+        {
+            var header = await DeserializeHeaderAsync(stream, cancellationToken).ConfigureAwait(false);
+
+            var items = new List<EnvelopeItem>();
+            while (stream.Position < stream.Length)
+            {
+                var item = await EnvelopeItem.DeserializeAsync(stream, cancellationToken).ConfigureAwait(false);
+                items.Add(item);
+            }
 
             return new Envelope(header, items);
         }