+++ src/Sentry/Protocol/EnvelopeItem.cs
@@ -1,3 +1,4 @@
+using System;
 using System.Collections.Generic;
 using System.IO;
 using System.Text;
@@ -11,9 +12,11 @@ namespace Sentry.Protocol
     /// <summary>
     /// Envelope item.
     /// </summary>
-    public class EnvelopeItem : ISerializable
+    public class EnvelopeItem : IDisposable, ISerializable
     {
         private const string TypeKey = "type";
+        private const string TypeValueEvent = "event";
+        private const string TypeValueUserReport = "user_report";
         private const string LengthKey = "length";
         private const string FileNameKey = "file_name";
 
@@ -47,16 +50,15 @@ namespace Sentry.Protocol
         public long? TryGetLength() =>
             Header.GetValueOrDefault(LengthKey) switch
             {
-                long x => x,
-                int x => x,
-                _ => null
+                null => null,
+                var value => Convert.ToInt64(value) // can be int, long, or another numeric type
             };
 
         private async Task<MemoryStream> BufferPayloadAsync(CancellationToken cancellationToken = default)
         {
             var buffer = new MemoryStream();
             await Payload.SerializeAsync(buffer, cancellationToken).ConfigureAwait(false);
-            _ = buffer.Seek(0, SeekOrigin.Begin);
+            buffer.Seek(0, SeekOrigin.Begin);
 
             return buffer;
         }
@@ -69,7 +71,7 @@ namespace Sentry.Protocol
             {
                 // Header
                 await Json.SerializeToStreamAsync(Header, stream, cancellationToken).ConfigureAwait(false);
-                stream.WriteByte((byte)'\n');
+                await stream.WriteByteAsync((byte)'\n', cancellationToken).ConfigureAwait(false);
 
                 // Payload
                 await Payload.SerializeAsync(stream, cancellationToken).ConfigureAwait(false);
@@ -85,12 +87,15 @@ namespace Sentry.Protocol
                 var headerData = Json.SerializeToByteArray(headerWithLength);
 
                 await stream.WriteAsync(headerData, cancellationToken).ConfigureAwait(false);
-                stream.WriteByte((byte)'\n');
+                await stream.WriteByteAsync((byte)'\n', cancellationToken).ConfigureAwait(false);
 
                 // Payload
                 await payloadBuffer.CopyToAsync(stream, cancellationToken).ConfigureAwait(false);
             }
         }
+
+        /// <inheritdoc />
+        public void Dispose() => (Payload as IDisposable)?.Dispose();
 
         /// <summary>
         /// Creates an envelope item from file.
@@ -137,7 +142,7 @@ namespace Sentry.Protocol
         {
             var header = new Dictionary<string, object>
             {
-                [TypeKey] = "event"
+                [TypeKey] = TypeValueEvent
             };
 
             return new EnvelopeItem(header, @event);
@@ -150,10 +155,103 @@ namespace Sentry.Protocol
         {
             var header = new Dictionary<string, object>
             {
-                [TypeKey] = "user_report"
+                [TypeKey] = TypeValueUserReport
             };
 
             return new EnvelopeItem(header, sentryUserFeedback);
+        }
+
+        private static async Task<IReadOnlyDictionary<string, object>> DeserializeHeaderAsync(
+            Stream stream,
+            CancellationToken cancellationToken = default)
+        {
+            var buffer = new List<byte>();
+
+            var prevByte = default(int);
+            await foreach (var curByte in stream.ReadAllBytesAsync(cancellationToken))
+            {
+                // Break if found an unescaped newline
+                if (curByte == '\n' && prevByte != '\\')
+                {
+                    break;
+                }
+
+                buffer.Add(curByte);
+                prevByte = curByte;
+            }
+
+            return
+                Json.DeserializeFromByteArray<Dictionary<string, object>?>(buffer.ToArray())
+                ?? throw new InvalidOperationException("Envelope item header is malformed.");
+        }
+
+        private static async Task<ISerializable> DeserializePayloadAsync(
+            Stream stream,
+            IReadOnlyDictionary<string, object> header,
+            CancellationToken cancellationToken = default)
+        {
+            var payloadLength = header.GetValueOrDefault(LengthKey) switch
+            {
+                null => (long?)null,
+                var value => Convert.ToInt64(value)
+            };
+
+            var payloadType = header.GetValueOrDefault(TypeKey) as string;
+
+            // Event
+            if (string.Equals(payloadType, TypeValueEvent, StringComparison.OrdinalIgnoreCase))
+            {
+                var bufferLength = (int)(payloadLength ?? stream.Length);
+                var buffer = await stream.ReadByteChunkAsync(bufferLength, cancellationToken).ConfigureAwait(false);
+
+                return Json.DeserializeFromByteArray<SentryEvent>(buffer);
+            }
+
+            // User report
+            if (string.Equals(payloadType, TypeValueUserReport, StringComparison.OrdinalIgnoreCase))
+            {
+                var bufferLength = (int)(payloadLength ?? stream.Length);
+                var buffer = await stream.ReadByteChunkAsync(bufferLength, cancellationToken).ConfigureAwait(false);
+
+                return Json.DeserializeFromByteArray<UserFeedback>(buffer);
+            }
+
+            // Arbitrary payload
+            if (payloadLength != null)
+            {
+                stream.Seek(payloadLength.Value, SeekOrigin.Current);
+            }
+            else
+            {
+                stream.Seek(0, SeekOrigin.End);
+            }
+
+            var payloadStream = new PartialStream(stream, stream.Position, payloadLength);
+
+            return new StreamSerializable(payloadStream);
+        }
+
+        /// <summary>
+        /// Deserializes envelope item from stream.
+        /// </summary>
+        public static async Task<EnvelopeItem> DeserializeAsync(
+            Stream stream,
+            CancellationToken cancellationToken = default)
+        {
+            var header = await DeserializeHeaderAsync(stream, cancellationToken).ConfigureAwait(false);
+            var payload = await DeserializePayloadAsync(stream, header, cancellationToken).ConfigureAwait(false);
+
+            // Swallow trailing newlines (some envelopes may have them after payloads)
+            await foreach (var curByte in stream.ReadAllBytesAsync(cancellationToken))
+            {
+                if (curByte != '\n')
+                {
+                    stream.Position--;
+                    break;
+                }
+            }
+
+            return new EnvelopeItem(header, payload);
         }
     }
 }